// src/controllers/media-contents/media-contents.interface.ts
export interface MediaContentsOptionInterface {
  mediaType?: string;
  url?: string;
  publicId?: string;
  userId?: number;
  format?: string;
  fileName?: string;
  path?: string;
  urlBlur?: string;
  urlTiny?: string;
  urlOptimize?: string;
  urlBackground?: string;
  resourceType?: string;
  metadata?: string;
}